Improve code readability
Be consistent in the code that we are covering with a next step
as I will probably want to also be able
to add upcoming or don't forget to dos
at some point in the future. And this will
keep the code more readable as to what
is really happening.