Remove do in a focus period choice
The motivation behind this change is because I am getting
an unimplemented panic when I try to look at the bullet list
for an item because the function get_estimated_focus_periods
is not implemented when deciding if the menu choice to give the
choice to estimate focus periods should be shown.

The change is to just always show this option for action and goal types.