Rename to Action & Goal mostly in DB layer
I have mostly switched to the Action & Goal terminology
in the UI. This change renames the types in the DB layer.

I am also setting it so a goal is either an aspirational hope or a
tangible milestone so the data layer can store this separation.
This is being done to avoid from needing to do a data storage
layer upgrade again later when I bring this change to the UI.

I have also done some renaming to Goal in some places in the UI
that have not been switched yet but that is mostly in UI that
is not really used much anymore.