Remove cover with milestone...
from mentally resident menu.

The motivation behind this change is I am working to change the
fields of the SurrealItem type and rather than update
this code I am going to just remove it in this PR
as it is not a used feature and this area needs to be reworked.