Prompt for enter list reason
When selecting On Deck or Mentally Resident ask the user what the enter list reason should be which can be a DateTime which what was an option before or immediately or highest uncovered. If HighestUncovered is selected then not implemented will be ran into but this upgrades the data layer to store the data required for this upcoming further change.