Towards implementing ItemStatus for lap_count and is_snoozed
This partially implements lap_count and is_snoozed along with unit tests. There is currently a failed Unit Test that requires making changes to other parts of OnPurpose so the proper data is available so I'm doing this check-in to save this much of the work before expanding to ensure that finished items are available.