Make it optional to store bullet list stats
This is because there are places where I want to log time spent when I don't have this data available.

Note that I am not upgrading the old data I am just making this change and then future data will be logged. Otherwise the old data will just be not available as in Option::None.