Fix bug so highest uncovered works properly
The issue was that when you selected highest uncovered it was not doing a recursive search so if you had children items on the higher then that would not be picked up and it would keep giving a new higher item. We now recursively search the tree so the nodes are what counts.