Add support for setting relationships during staging
The code is largely untested as there is no way to yield to the parent item for the lap count. Making this more of a point in time and in progress check-in, but I've made a lot of changes that I want to stay with and I have tested it to work so I'm going to check-in.