Rethinking how to store data to deal with a new priority model.
This code does not compile but I am checking it in to a new branch becuase I am using this to work out what changes need to be done in the main branch in order to prepare for this change. Which in summary is to do away with the SurrealStaging type, but I will need to do that migration piece by piece in the main branch as I work out what needs to be done in this branch.