Bug Fix: Can go to parent again
There was an issue with the PartialEq implementation, really we should just be checking the RecordId of the SurrealDB record rather than a full check and this causes that behavior so we no longer need to manually do that.

This also removes some dead code that showed up during this work.