Show selected item and print up to 8 smaller items
Print out the selected item to make it easier to know what is selected when navigating through multiple menu options. Also make it so at least some of the completed items is shown each time rather than going to just a number. Up to 8 completed are shown and then the rest are just shown as a number.