Prompt user to see a kind to motivation if that is not set.
This probably only impacts those who have been using On Purpose for a while  as today when you set a motivation you set a kind of it being core work or non-core work. This is being done because I am working on adding a mode filtering feature of filtering to a certain mode so having this value set is required so you can select core or non-core work.