Improve performance
Use a different allocator that seems to give about a 15% performance improvement when running the calculate data portion of the code. I believe the reason for this is because of all of the visited allocations that happen when traversing the tree. At some point this might need to be reworked in some way so it is not a vector that is copied a ton but for now this seems to be enough.