Importance & urgency in time spent & reflection
Log importance and urgency in time spent and display this in reflection

Now if an item is both important and urgent both icons are shown.

In the main do now list all icons are shown for all urgencies if it contains a most important item in the select from list item.