Updates to improve usability
Remove [NEEDS NEXT STEP] if it is a project with no children.
Update some of the unicode icons to make things more pleasant.
Say "Do Now" for items to work on.
If an incorrect date time is entered when saying when to work on something next then prompt again.
Add more spaces between different prompts in the UI