Add support for Events
An event is something that can happen to uncover additional items that can now be done. For example you can say that I need to wait until I get home before I can do this item and then
a wait until you are home event needs to be triggered. This shows up in the menu and you can trigger the event or see all of the items waiting on this event.