Stable order to priorities, oldest to latest
Sometimes when using the priority feature I was running into an issue were when I would set a new priority because it was then preventing other priorities from being applied I was then seeing other items show up and this was wasting time setting more priorities. By causing a stable ordering and by setting that ordering to be from oldest to newest the priority order that things are applied should now be stable and cause the older priorities to still be applied first making it so new things don't show up by surprise.